3.x merge - PH_STRING conversion

+++ ProcessHacker/netprv.c
@@ -380,7 +380,7 @@ PPH_STRING PhGetHostNameFromAddress(
         address,
         length,
         hostName->Buffer,
-        hostName->Length / 2 + 1,
+        (ULONG)hostName->Length / 2 + 1,
         NULL,
         0,
         NI_NAMEREQD
@@ -394,7 +394,7 @@ PPH_STRING PhGetHostNameFromAddress(
             address,
             length,
             hostName->Buffer,
-            hostName->Length / 2 + 1,
+            (ULONG)hostName->Length / 2 + 1,
             NULL,
             0,
             NI_NAMEREQD
